Disable LLVM_ENABLE_THREADS explicitly in configs.
On DragonFly in base system bootstrapping tools it is preferable that they
should be as simple as possible to ease up future updates. Having tools
depending on threading is considered harmful and should be avoided.

Sadly just defining LLVM_ENABLE_THREADS to 0 is not enough and some patching
is still needed in few contrib sources before -lpthread can be dropped from
LDFLAGS.