prune: Some of MS extensions to c++.
This is needed for next parts, code is very fragile and sprayed all over the
place with certain bugs in checking logic that leaks into other invoke modes.
Especially in asm and sema sources that were doing certain unwanted things.
This includes removal of Borland extensions as well because of complexities.