prune: ObjC Runtime library support.
Initial effort was just to kill the implicit passing of -fobjc-runtime=gcc
in clang -cc1 args, however when doing that it uncovered the sorry state of
whole ObjC extension to C language implementation. I could stop here, but
ObjC bits are dangerously dangling in the wind so naturally it makes sense to
go further and clean up clang's codebase even more.
Nothing in base system depends on such extensions.