cleanup: Fix warnings for CCVER=clangb WARNS=2
It is strange why clang's -Wunused-private-field gets triggered on its own code
that is comming from -Wall, behaviour is identical to devel/llvm39 port so for
the time being allow to emit these spurious diagnostics.

Also for -O2 compilations enforce -fno-strict aliasing, there are issues in
clang source codes but be generous and assume the same is needed for llvm too.