build: For now just disable LLVMgold.so building.
Turns out it is very incompatible with lto objects from gcc <-> clang.
Using any of compiler final linkage is impossible with both flaviours
of LTO objects. Also given that it is useful only for specific use cases
it is reasonable to disable it by default to save build time.
LLVMgold.so takes about 1/3 of build time to build its own -fpic library.